sc build updates.
DynamicForm : Don't show required styling for non-editable elements
DataSource : Added the ability to dynamically set clientCustom in transformRequest()
Fixed a bug where a wsdl:part referring to an element of simpleType in an
imported namespace would not find the right namespace.
For a wsdl:part of simpleType, automatically accept a few different possible
ways of trying to specify the data.

various other enhancements / fixes