Reverted r2362
That change caused a regression of gridComponents:com/smartgwt/sample/showcase/client/grid/appearance/GridComponentsSample #1
There, a ResultSet "properties" object was created for the purpose of adding
a DataChangedHandler to a SmartClient ResultSet instance to be created. However,
the code did not want the SC ResultSet object to be auto-created as a result of
adding the DataChangedHandler.

This commit adds a new API to ResultSet, ResultSet#ensureCreated(), which can
be called for standalone usage of a ResultSet.

[dt]